Remove i386 specific code
