--- i686-rxinu.json
@@ -1,14 +1,0 @@
-{
-  "llvm-target": "i686-unknown-none",
-  "data-layout": "e-m:e-p:32:32-f64:32:64-f80:32-n8:16:32-S128",
-  "target-endian": "little",
-  "target-pointer-width": "32",
-  "target-c-int-width": "32",
-  "arch": "x86",
-  "os": "none",
-  "features": "-mmx,-sse,+soft-float",
-  "disable-redzone": true,
-  "linker-flavor": "ld.lld",
-  "panic-strategy": "abort",
-  "executables": true
-}
