+++ src/arch/mod.rs
@@ -1,6 +1,6 @@
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+#[cfg(target_arch = "x86_64")]
 #[macro_use]
-pub mod x86;
+pub mod x86_64;
 
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-pub use self::x86::*;
+#[cfg(target_arch = "x86_64")]
+pub use self::x86_64::*;
