+++ src/arch/x86_64/context_switch.asm
@@ -4,7 +4,6 @@
 .intel_syntax noprefix
 # Context {
 #   0x0: flags
-#   0x8: cr3
 #   0x8: rbx
 #   0x10: r12
 #   0x18: r13
@@ -22,24 +21,24 @@ x86_64_context_switch:
     pop qword ptr [rdi] # save rflags into prev.flags
     # Rust inline assembly error: invalid operand
     #mov [rdi+0x8], cr3  # save rbx
-    mov [rdi+0x10], rbx  # save rbx
-    mov [rdi+0x18], r12 # save r12
-    mov [rdi+0x20], r13 # save r13
-    mov [rdi+0x28], r14 # save r14
-    mov [rdi+0x30], r15 # save r15
-    mov [rdi+0x38], rbp # save rbp
+    mov [rdi+0x08], rbx  # save rbx
+    mov [rdi+0x10], r12 # save r12
+    mov [rdi+0x18], r13 # save r13
+    mov [rdi+0x20], r14 # save r14
+    mov [rdi+0x28], r15 # save r15
+    mov [rdi+0x30], rbp # save rbp
 
     # Swap the stack pointers
-    mov [rdi+0x40], rsp # save rsp
-    mov rsp, [rsi+0x40] # set rsp
+    mov [rdi+0x38], rsp # save rsp
+    mov rsp, [rsi+0x38] # set rsp
 
     # Switch to the next context
-    mov rbp, [rsi+0x38] # set rbp
-    mov r15, [rsi+0x30] # set r15
-    mov r14, [rsi+0x28] # set r14
-    mov r13, [rsi+0x20] # set r13
-    mov r12, [rsi+0x18] # set r12
-    mov rbx, [rsi+0x10]  # set rbx
+    mov rbp, [rsi+0x30] # set rbp
+    mov r15, [rsi+0x28] # set r15
+    mov r14, [rsi+0x20] # set r14
+    mov r13, [rsi+0x18] # set r13
+    mov r12, [rsi+0x10] # set r12
+    mov rbx, [rsi+0x08]  # set rbx
     # Rust inline assembly error: invalid operand
     #mov cr3, [rsi+0x8]  # set rbx
     push [rsi] # set rflags
